requirements.txt
flask
fpdf2==2.7.9

